Refactor admin dashboard components
This commit refactors the admin dashboard, improving code organization and component structure. It extracts the ProblemForm component, introduces a Pinia store for problem management, and improves image handling operations.

The main changes include:
1. Extraction of ProblemForm into a separate component for better maintainability.
2. Introduction of a Pinia store (problemsStore) to centralize problem-related operations.
3. Refactoring of image management logic into a separate composable (useImageManagement).
4. Simplification of the AdminDashboard component by delegating responsibilities to child components and the store.
