+++ src/main.js
@@ -1,8 +1,12 @@
 import {createApp} from 'vue'
+import {createPinia} from 'pinia'
 import App from './App.vue'
 import router from './router'
 import 'highlight.js/styles/github.css'
 
 const app = createApp(App)
+const pinia = createPinia()
+
+app.use(pinia)
 app.use(router)
 app.mount('#app')
